Camera.hpp moved
