+++ examples_tests/common/Camera.hpp
